Create coinchange.py
You are given an integer array coins representing coins of different denominations and an integer amount representing a total amount of money.Return the number of combinations that make up that amount. If that amount of money cannot be made up by any combination of the coins, return 0.
//Approach//
Initially we have to declare an array with 1 more element than the amount.
since we dont need coins for zero set it to zero.
take each coin and find no of coins required to generate the corresponding index.