Create Uniquepath2.py
You are given an m x n integer array grid. There is a robot initially located at the top-left corner (i.e., grid[0][0]). The robot tries to move to the bottom-right corner (i.e., grid[m - 1][n - 1]). The robot can only move either down or right at any point in time.

An obstacle and space are marked as 1 or 0 respectively in grid. A path that the robot takes cannot include any square that is an obstacle.

Return the number of possible unique paths that the robot can take to reach the bottom-right corner.

//Approach and Intuition//

first find no. of columns and rows 
create a matrix to store the no. of ways in which we could reach every point in the grid.
Since it can only move right and downwards set initial value.
store zero when a obstacle is found.
Add number of paths from and top and left